MINOR: Move JaasModule into JaasUtils
