+++ test-common/test-common-runtime/src/main/java/org/apache/kafka/common/test/KafkaClusterTestKit.java
@@ -260,7 +260,7 @@ public class KafkaClusterTestKit implements AutoCloseable {
                 File file = JaasUtils.writeJaasContextsToFile(Set.of(
                     new JaasUtils.JaasSection(JaasUtils.KAFKA_SERVER_CONTEXT_NAME,
                         List.of(
-                            JaasModule.plainLoginModule(
+                            JaasUtils.JaasModule.plainLoginModule(
                                 JaasUtils.KAFKA_PLAIN_ADMIN, 
                                 JaasUtils.KAFKA_PLAIN_ADMIN_PASSWORD,
                                 true,
